+++ packages/adapter-interfaces/src/types/Code.ts
@@ -10,7 +10,7 @@ export const codeTypeSchema = z.enum([
 ]);
 export type CodeType = z.infer<typeof codeTypeSchema>;
 
-export const codeSchema = z.object({
+export const codeInsertSchema = z.object({
   code_id: z.string().openapi({
     description:
       "The code that will be used in for instance an email verification flow",
@@ -19,9 +19,15 @@ export const codeSchema = z.object({
     description: "The id of the login session that the code is connected to",
   }),
   code_type: codeTypeSchema,
-  created_at: z.string(),
   expires_at: z.string(),
   used_at: z.string().optional(),
+});
+
+export type CodeInsert = z.infer<typeof codeInsertSchema>;
+
+export const codeSchema = z.object({
+  ...codeInsertSchema.shape,
+  created_at: z.string(),
 });
 
 export type Code = z.infer<typeof codeSchema>;