+++ packages/adapter-interfaces/src/types/Login.ts
@@ -3,13 +3,10 @@ import { authParamsSchema } from "./AuthParams";
 
 export const loginInsertSchema = z
   .object({
-    login_id: z.string().openapi({
-      description: "This is is used as the state in the universal login",
-    }),
     auth0Client: z.string().optional(),
     authParams: authParamsSchema,
     expires_at: z.string(),
-    deleted_at: z.string(),
+    deleted_at: z.string().optional(),
     ip: z.string().optional(),
   })
   .openapi({
@@ -20,6 +17,9 @@ export type LoginInsert = z.infer<typeof loginInsertSchema>;
 
 export const loginSchema = z.object({
   ...loginInsertSchema.shape,
+  login_id: z.string().openapi({
+    description: "This is is used as the state in the universal login",
+  }),
   created_at: z.string(),
   updated_at: z.string(),
 });