+++ packages/kysely/src/codes/create.ts
@@ -1,14 +1,12 @@
-import { Code } from "@authhero/adapter-interfaces";
+import { Code, CodeInsert } from "@authhero/adapter-interfaces";
 import { Kysely } from "kysely";
 import { Database } from "../db";
 
 export function create(db: Kysely<Database>) {
-  return async (tenant_id: string, code: Code) => {
-    const createdCode = {
+  return async (tenant_id: string, code: CodeInsert) => {
+    const createdCode: Code = {
       ...code,
-      tenant_id,
       created_at: new Date().toISOString(),
-      updated_at: new Date().toISOString(),
     };
 
     await db