+++ packages/kysely/src/logins/create.ts
@@ -1,4 +1,4 @@
-import { LoginInsert } from "@authhero/adapter-interfaces";
+import { Login, LoginInsert } from "@authhero/adapter-interfaces";
 import { Kysely } from "kysely";
 import { nanoid } from "nanoid";
 import { Database } from "../db";
@@ -6,8 +6,8 @@ import { flattenObject } from "../flattten";
 
 export function create(db: Kysely<Database>) {
   return async (tenant_id: string, login: LoginInsert) => {
-    const createdLogin = {
-      id: nanoid(),
+    const createdLogin: Login = {
+      login_id: nanoid(),
       ...login,
       created_at: new Date().toISOString(),
       updated_at: new Date().toISOString(),