+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/kysely-adapter
 
+## 0.5.5
+
+### Patch Changes
+
+- Updated the types for logins and fixed the packaging for authhero
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.10.0
+
 ## 0.5.4
 
 ### Patch Changes